tags-input: expose types
